requirements.txt
appdirs==1.4.3
attrs==16.3.0
packaging==16.8
py==1.4.32
pyparsing==2.2.0
pytest==3.0.6
six==1.10.0
